feedback tweaks
