+++ lib/appcore.js
@@ -173,29 +173,7 @@ var proto = {
             delegate.requestStart(app);
         }
 
-        // handle forms with formidable
-        app.use(function(req, res, next) {
-            var formidable = require('formidable'), form = new formidable.IncomingForm();
-            form.parse(req, function(err, fields, files) {
-                req.fields = fields; // pass along form fields
-                req.files = files;   // pass along files
-
-                // remove tmp files after request finishes
-                var cleanup = function() {
-                    var fs = fs || require('fs'), i;
-                    if (files) {
-                        for (i in files) {
-                            fs.unlink(files[i].path);
-                        }
-                    }
-                };
-                res.on('finish', cleanup);
-                res.on('close', cleanup);
-                res.on('error', cleanup);
-                next();
-            });
-        });
-
+        app.use(kraken.formidable(config.formidableSettings));
         app.use(express.cookieParser(config.session.secret));
         app.use(kraken.session(config.session));
         app.use(kraken.appsec(config.appsec));
