+++ lib/middleware/formidable.js
@@ -19,25 +19,30 @@
 /***@@@ END LICENSE @@@***/
 'use strict';
 
-var error = require('./error');
 
+module.exports = function (settings) {
+    var settings = typeof settings === 'object' ? settings : {};
 
-exports = module.exports = {
+    return function(req, res, next) {
+        var formidable = require('formidable'), form = new formidable.IncomingForm(settings);
+        form.parse(req, function(err, fields, files) {
+            req.fields = fields; // pass along form fields
+            req.files = files;   // pass along files
 
-    logger: require('./logger'),
-
-    session: require('./session'),
-
-    appsec: require('lusca'),
-
-    formidable: require('./formidable'),
-
-    compiler: require('./compiler'),
-
-    fileNotFound: error.fileNotFound,
-
-    serverError: error.serverError,
-
-    errorHandler: error.defaultHandler
+            // remove tmp files after request finishes
+            var cleanup = function() {
+                var fs = fs || require('fs'), i;
+                if (files) {
+                    for (i in files) {
+                        fs.unlink(files[i].path);
+                    }
+                }
+            };
+            res.on('finish', cleanup);
+            res.on('close', cleanup);
+            res.on('error', cleanup);
+            next();
+        });
+    };
 
 };
